+++ src/Helpers/TorrentUtils.php
@@ -2,10 +2,62 @@
 
 namespace Popstas\Transmission\Console\Helpers;
 
+use InfluxDB;
 use Martial\Transmission\API\Argument\Torrent;
 
 class TorrentUtils
 {
+    /**
+     * @param array $torrent
+     * @param string $transmissionHost
+     * @param array $lastPoint
+     * @return InfluxDB\Point
+     */
+    public static function buildPoint(array $torrent, $transmissionHost, array $lastPoint)
+    {
+        $age = TorrentUtils::getTorrentAge($torrent);
+
+        $tagsData = [
+            'host'             => $transmissionHost,
+            'torrent_name'     => $torrent[Torrent\Get::NAME],
+        ];
+
+        $uploadedDerivative = count($lastPoint) && $torrent[Torrent\Get::UPLOAD_EVER] - $lastPoint['last'] >= 0 ?
+            $torrent[Torrent\Get::UPLOAD_EVER] - $lastPoint['last'] : $torrent[Torrent\Get::UPLOAD_EVER];
+
+        $fieldsData = [
+            'uploaded_derivative' => $uploadedDerivative,
+            'downloaded'          => $torrent[Torrent\Get::TOTAL_SIZE],
+            'age'                 => $age,
+            'uploaded_per_day'    => $age ? $torrent[Torrent\Get::UPLOAD_EVER] / $age * 86400 : 0,
+        ];
+
+        return new InfluxDB\Point(
+            'uploaded',
+            $torrent[Torrent\Get::UPLOAD_EVER],
+            $tagsData,
+            $fieldsData,
+            time()
+        );
+    }
+
+    public static function getLastPoint(array $torrent, $transmissionHost, InfluxDB\Database $database)
+    {
+        $torrentName = $torrent[Torrent\Get::NAME];
+        $queryBuilder = $database->getQueryBuilder();
+        $results = $queryBuilder
+            ->last('value')
+            ->from('uploaded')
+            ->where([
+                "host='$transmissionHost'",
+                "torrent_name='$torrentName'"
+            ])
+            ->getResultSet()
+            ->getPoints();
+
+        return count($results) ? $results[0] : [];
+    }
+
     /**
      * @param $torrent
      * @return int seconds from torrent finish download
@@ -25,6 +77,7 @@ class TorrentUtils
 
     public static function getSizeInGb($sizeInBytes, $round = 2)
     {
-        return round($sizeInBytes / 1024 / 1024 / 1024, $round);
+        // 1024 not equals transmission value
+        return round($sizeInBytes / 1000 / 1000 / 1000, $round);
     }
 }